# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Service - API RESTFULL",
    "description": "API RESTFULL sobre productos y pedidos",
    "version": "2.1.0"
  },
  "servers": [{"url": "http://localhost:3000"}],
  "tags": [
    {"name": "Auth"},
    {"name": "Productos v1"},
    {"name": "Pedidos v1"},
    {"name": "Productos v2"}
  ],
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Autenticar y obtener JWT",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginRequest"}}}},
        "responses": {
          "200": {"description": "Token generado", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginResponse"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "422": {"$ref": "#/components/responses/Validation"}
        }
      }
    },
    "/api/v1/productos": {
      "get": {
        "tags": ["Productos v1"],
        "summary": "Listar productos",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "page", "in": "query", "schema": {"type": "integer", "minimum": 1, "default": 1}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 50, "default": 10}},
          {"name": "search", "in": "query", "schema": {"type": "string"}}
        ],
        "responses": {"200": {"description": "Lista de productos"}, "401": {"$ref": "#/components/responses/Unauthorized"}}
      },
      "post": {
        "tags": ["Productos v1"],
        "summary": "Crear producto",
        "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductInput"}}}},
        "responses": {
          "201": {"description": "Producto creado", "headers": {"Location": {"schema": {"type": "string"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "422": {"$ref": "#/components/responses/Validation"}
        }
      }
    },
    "/api/v1/productos/{id}": {
      "parameters": [{"$ref": "#/components/parameters/ProductId"}],
      "get": {
        "tags": ["Productos v1"], "summary": "Obtener producto", "security": [{"bearerAuth": []}],
        "responses": {"200": {"description": "Producto"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "404": {"$ref": "#/components/responses/NotFound"}}
      },
      "put": {
        "tags": ["Productos v1"], "summary": "Reemplazar producto", "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductInput"}}}},
        "responses": {"200": {"description": "Actualizado"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "404": {"$ref": "#/components/responses/NotFound"}, "422": {"$ref": "#/components/responses/Validation"}}
      },
      "patch": {
        "tags": ["Productos v1"], "summary": "Modificar producto", "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductPatch"}}}},
        "responses": {"200": {"description": "Actualizado"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "404": {"$ref": "#/components/responses/NotFound"}, "422": {"$ref": "#/components/responses/Validation"}}
      },
      "delete": {
        "tags": ["Productos v1"], "summary": "Desactivar producto", "security": [{"bearerAuth": []}],
        "responses": {"204": {"description": "Sin contenido"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "404": {"$ref": "#/components/responses/NotFound"}}
      }
    },
    "/api/v1/pedidos": {
      "get": {"tags": ["Pedidos v1"], "summary": "Listar pedidos", "security": [{"bearerAuth": []}], "responses": {"200": {"description": "Pedidos"}, "401": {"$ref": "#/components/responses/Unauthorized"}}},
      "post": {
        "tags": ["Pedidos v1"], "summary": "Crear pedido", "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OrderInput"}}}},
        "responses": {"201": {"description": "Pedido creado", "headers": {"Location": {"schema": {"type": "string"}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}, "422": {"$ref": "#/components/responses/Validation"}}
      }
    },
    "/api/v1/pedidos/{id}": {
      "parameters": [{"$ref": "#/components/parameters/OrderId"}],
      "get": {"tags": ["Pedidos v1"], "summary": "Obtener pedido", "security": [{"bearerAuth": []}], "responses": {"200": {"description": "Pedido"}, "404": {"$ref": "#/components/responses/NotFound"}}},
      "put": {"tags": ["Pedidos v1"], "summary": "Actualizar estado", "security": [{"bearerAuth": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OrderStateInput"}}}}, "responses": {"200": {"description": "Actualizado"}, "422": {"$ref": "#/components/responses/Validation"}}},
      "patch": {"tags": ["Pedidos v1"], "summary": "Modificar estado", "security": [{"bearerAuth": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OrderStateInput"}}}}, "responses": {"200": {"description": "Actualizado"}, "422": {"$ref": "#/components/responses/Validation"}}},
      "delete": {"tags": ["Pedidos v1"], "summary": "Cancelar pedido", "security": [{"bearerAuth": []}], "responses": {"204": {"description": "Sin contenido"}, "404": {"$ref": "#/components/responses/NotFound"}}}
    },
    "/api/v1/pedidos/{id}/items": {
      "get": {"tags": ["Pedidos v1"], "summary": "Consultar ítems del pedido", "security": [{"bearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/OrderId"}], "responses": {"200": {"description": "Ítems"}, "404": {"$ref": "#/components/responses/NotFound"}}}
    },
    "/api/v2/productos": {
      "get": {
        "tags": ["Productos v2"], "summary": "Listar productos v2 con Rate Limiting", "security": [{"bearerAuth": []}],
        "responses": {"200": {"description": "Lista v2"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "429": {"$ref": "#/components/responses/RateLimit"}}
      },
      "post": {
        "tags": ["Productos v2"], "summary": "Crear producto v2", "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductInput"}}}},
        "responses": {"201": {"description": "Creado", "headers": {"Location": {"schema": {"type": "string"}}, "X-RateLimit-Limit": {"schema": {"type": "integer"}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}, "422": {"$ref": "#/components/responses/Validation"}, "429": {"$ref": "#/components/responses/RateLimit"}}
      }
    },
    "/api/v2/productos/{id}": {
      "parameters": [{"$ref": "#/components/parameters/ProductId"}],
      "get": {"tags": ["Productos v2"], "summary": "Obtener producto v2", "security": [{"bearerAuth": []}], "responses": {"200": {"description": "Producto"}, "404": {"$ref": "#/components/responses/NotFound"}, "429": {"$ref": "#/components/responses/RateLimit"}}},
      "put": {"tags": ["Productos v2"], "summary": "Reemplazar producto v2", "security": [{"bearerAuth": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductInput"}}}}, "responses": {"200": {"description": "Actualizado"}, "422": {"$ref": "#/components/responses/Validation"}, "429": {"$ref": "#/components/responses/RateLimit"}}},
      "patch": {"tags": ["Productos v2"], "summary": "Modificar producto v2", "security": [{"bearerAuth": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductPatch"}}}}, "responses": {"200": {"description": "Actualizado"}, "422": {"$ref": "#/components/responses/Validation"}, "429": {"$ref": "#/components/responses/RateLimit"}}},
      "delete": {"tags": ["Productos v2"], "summary": "Desactivar producto v2", "security": [{"bearerAuth": []}], "responses": {"204": {"description": "Sin contenido"}, "429": {"$ref": "#/components/responses/RateLimit"}}}
    }
  },
  "components": {
    "securitySchemes": {"bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}},
    "parameters": {
      "ProductId": {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "minimum": 1}},
      "OrderId": {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "minimum": 1}}
    },
    "schemas": {
      "LoginRequest": {"type": "object", "required": ["email", "password"], "properties": {"email": {"type": "string", "format": "email", "example": "admin@ecommerce.ec"}, "password": {"type": "string", "format": "password", "example": "Admin123*"}}},
      "LoginResponse": {"type": "object", "properties": {"token": {"type": "string"}, "tokenType": {"type": "string", "example": "Bearer"}, "expiresIn": {"type": "string", "example": "2h"}}},
      "ProductInput": {"type": "object", "required": ["nombre", "precio", "stock", "activo"], "properties": {"nombre": {"type": "string", "example": "Monitor 24 pulgadas"}, "descripcion": {"type": "string", "example": "Monitor Full HD"}, "precio": {"type": "number", "minimum": 0, "example": 149.99}, "stock": {"type": "integer", "minimum": 0, "example": 10}, "activo": {"type": "boolean", "example": true}}},
      "ProductPatch": {"type": "object", "properties": {"nombre": {"type": "string"}, "descripcion": {"type": "string"}, "precio": {"type": "number", "minimum": 0}, "stock": {"type": "integer", "minimum": 0}, "activo": {"type": "boolean"}}},
      "OrderInput": {"type": "object", "required": ["items"], "properties": {"items": {"type": "array", "minItems": 1, "items": {"type": "object", "required": ["productId", "cantidad"], "properties": {"productId": {"type": "integer", "minimum": 1}, "cantidad": {"type": "integer", "minimum": 1}}}}}},
      "OrderStateInput": {"type": "object", "required": ["estado"], "properties": {"estado": {"type": "string", "enum": ["PENDIENTE", "CONFIRMADO", "ENVIADO", "ENTREGADO", "CANCELADO"]}}},
      "Error": {"type": "object", "properties": {"error": {"type": "object", "properties": {"code": {"type": "string"}, "message": {"type": "string"}, "details": {"type": "array", "items": {"type": "string"}}}}}}
    },
    "responses": {
      "Unauthorized": {"description": "No autorizado", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "Validation": {"description": "Error de validación", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "NotFound": {"description": "Recurso no encontrado", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "RateLimit": {"description": "Too Many Requests", "headers": {"X-RateLimit-Limit": {"description": "Máximo de solicitudes por ventana", "schema": {"type": "integer", "example": 5}}, "Retry-After": {"description": "Segundos de espera", "schema": {"type": "integer", "example": 60}}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}
    }
  }
}